+++ tests/test_micc.py
@@ -173,6 +173,6 @@ if __name__ == "__main__":
     print(sys.version_info)
     the_test_you_want_to_debug = test_scenario_2
 
-    with micc.logging.log(print,f"__main__ running {the_test_you_want_to_debug}",'-*# finished #*-'):
+    with micc.logging_tools.log(print,f"__main__ running {the_test_you_want_to_debug}",'-*# finished #*-'):
         the_test_you_want_to_debug()
 # ==============================================================================